Contract - Optimaize duplicate code
