+++ components/modules/contracts/index.ts
@@ -1,2 +1,3 @@
 export * from './SelectCompanyModal';
 export * from './SelectEmployeeModal';
+export * from './ContractForm';
