+++ pages/contracts/create.tsx
@@ -1,35 +1,32 @@
 import { uuidv4 } from '@firebase/util';
-import { Button, DatePicker, Form, Input, InputNumber, message, Select, Space, Typography } from 'antd';
+import { message, Space, Typography } from 'antd';
 import { useForm } from 'antd/lib/form/Form';
 import { initializeApp } from 'firebase/app';
 import { getDownloadURL, getStorage, ref, uploadBytes } from 'firebase/storage';
 import getConfig from 'next/config';
-import Link from 'next/link';
 import { useRouter } from 'next/router';
 import { useState } from 'react';
 
 import { contractApi } from '~/api-clients/modules/contract-api';
-import { Seo, UploadFiles } from '~/components';
-import { SelectCompanyModal, SelectEmployeeModal } from '~/components/modules/contracts';
+import { Seo } from '~/components';
+import { ContractForm } from '~/components/modules/contracts';
 import { firebaseConfig } from '~/firebaseconfig';
-import { Regex } from '~/models/constants/Regex';
 import { AppRoutes } from '~/models/constants/Routes';
 import { Company } from '~/models/modules/companies';
-import { CreateForm, CreatePayload, SalaryType, Type } from '~/models/modules/contracts';
+import { CreatePayload, FormData, SalaryType, Type } from '~/models/modules/contracts';
 import { Employee } from '~/models/modules/employees';
 
 const { serverRuntimeConfig } = getConfig();
 
 export default function CreateContractPage() {
   const router = useRouter();
-  const { RangePicker } = DatePicker;
   const [form] = useForm();
   const [loading, setLoading] = useState<boolean>(false);
   const [file, setFile] = useState<any>(null);
   const [company, setCompany] = useState<Company | null>(null);
   const [employee, setEmployee] = useState<Employee | null>(null);
 
-  const onFinish = async (data: CreateForm) => {
+  const onFinish = async (data: FormData) => {
     setLoading(true);
     try {
       const app = initializeApp(firebaseConfig);
@@ -64,26 +61,6 @@ export default function CreateContractPage() {
     setLoading(false);
   };
 
-  const handleSelectCompany = (company: Company | null) => {
-    if (company) {
-      setCompany(company);
-      form.setFieldsValue({ companyId: company.companyId });
-    } else {
-      setCompany(null);
-      form.setFieldsValue({ companyId: '' });
-    }
-  };
-
-  const handleSelectEmployee = (employee: Employee | null) => {
-    if (employee) {
-      setEmployee(employee);
-      form.setFieldsValue({ employeeId: employee.employeeId });
-    } else {
-      setCompany(null);
-      form.setFieldsValue({ employeeId: '' });
-    }
-  };
-
   return (
     <>
       <Seo
@@ -98,174 +75,20 @@ export default function CreateContractPage() {
           <Typography.Title level={3}>Create new contract</Typography.Title>
         </section>
 
-        <Form
-          form={form}
-          labelCol={{ span: 8 }}
-          wrapperCol={{ span: 16 }}
-          style={{ maxWidth: 600 }}
-          initialValues={{ remember: true }}
-          onFinish={onFinish}
-          autoComplete="off"
-        >
-          <Space style={{ width: '100%' }} direction="vertical">
-            <Form.Item label="File" name="file" rules={[{ required: true, message: 'Please add contract file!' }]}>
-              <UploadFiles
-                file={file}
-                setFile={(file) => {
-                  if (file) {
-                    setFile(file);
-                    form.setFieldsValue({ file: file.name });
-                  } else {
-                    setFile(null);
-                    form.setFieldsValue({ file: '' });
-                  }
-                }}
-              />
-            </Form.Item>
-            <Form.Item
-              label="Apply date"
-              name="applyDate"
-              rules={[{ required: true, message: 'Please input apply date!' }]}
-            >
-              <RangePicker format={'DD/MM/YYYY'} />
-            </Form.Item>
-            <Form.Item label="Job title" name="job" rules={[{ required: true, message: 'Please input job title!' }]}>
-              <Input />
-            </Form.Item>
-            <Form.Item
-              label="Basic salary"
-              name="basicSalary"
-              rules={[{ required: true, message: 'Please input basic salary!' }]}
-            >
-              <InputNumber
-                formatter={(value) => `${value}`.replace(Regex.COMMAS_SEPARATED_NUMBER, ',')}
-                addonAfter="VNĐ"
-                min={0}
-                max={1000000000}
-                style={{ width: '100%' }}
-              />
-            </Form.Item>
-            <Form.Item
-              label="Salary type"
-              name="salaryType"
-              rules={[{ required: true, message: 'Please select salary type!' }]}
-            >
-              <Select
-                showSearch
-                placeholder="Select salary type"
-                allowClear
-                onClear={() => form.setFieldsValue({ salaryType: '' })}
-                onChange={(value) => form.setFieldsValue({ salaryType: value })}
-                filterOption={(input, option) => (option?.label ?? '').toLowerCase().includes(input.toLowerCase())}
-                options={[
-                  {
-                    value: SalaryType.Gross,
-                    label: 'Gross',
-                  },
-                  {
-                    value: SalaryType.Net,
-                    label: 'Net',
-                  },
-                ]}
-              />
-            </Form.Item>
-            <Form.Item label="BHXH" name="bhxh" rules={[{ required: true, message: 'Please input price of BHXH!' }]}>
-              <InputNumber
-                formatter={(value) => `${value}`.replace(Regex.COMMAS_SEPARATED_NUMBER, ',')}
-                addonAfter="VNĐ"
-                min={0}
-                max={1000000000}
-                style={{ width: '100%' }}
-              />
-            </Form.Item>
-            <Form.Item label="BHYT" name="bhyt" rules={[{ required: true, message: 'Please input price of BHYT!' }]}>
-              <InputNumber
-                formatter={(value) => `${value}`.replace(Regex.COMMAS_SEPARATED_NUMBER, ',')}
-                addonAfter="VNĐ"
-                min={0}
-                max={1000000000}
-                style={{ width: '100%' }}
-              />
-            </Form.Item>
-            <Form.Item label="BHTN" name="bhtn" rules={[{ required: true, message: 'Please input price of BHTN!' }]}>
-              <InputNumber
-                formatter={(value) => `${value}`.replace(Regex.COMMAS_SEPARATED_NUMBER, ',')}
-                addonAfter="VNĐ"
-                min={0}
-                max={1000000000}
-                style={{ width: '100%' }}
-              />
-            </Form.Item>
-            <Form.Item noStyle>
-              <Form.Item
-                label="Company"
-                name="companyId"
-                rules={[{ required: true, message: 'Please select a company!' }]}
-              >
-                <SelectCompanyModal setCompany={handleSelectCompany} companyName={company?.companyName || ''} />
-              </Form.Item>
-            </Form.Item>
-            <Form.Item
-              label="Company price"
-              name="companyPrice"
-              rules={[{ required: true, message: 'Please enter company price!' }]}
-            >
-              <InputNumber
-                formatter={(value) => `${value}`.replace(Regex.COMMAS_SEPARATED_NUMBER, ',')}
-                addonAfter="VNĐ"
-                min={0}
-                max={1000000000}
-                style={{ width: '100%' }}
-              />
-            </Form.Item>
-            <Form.Item
-              label="Employee"
-              name="employeeId"
-              rules={[{ required: true, message: 'Please select an employee!' }]}
-            >
-              <SelectEmployeeModal setEmployee={handleSelectEmployee} employeeName={employee?.name || ''} />
-            </Form.Item>
-            <Form.Item label="Type" name="type" rules={[{ required: true, message: 'Please select contract type!' }]}>
-              <Select
-                showSearch={true}
-                placeholder="Select contract type"
-                allowClear={true}
-                onClear={() => form.setFieldsValue({ type: '' })}
-                onChange={(value) => form.setFieldsValue({ type: value })}
-                filterOption={(input, option) => (option?.label ?? '').toLowerCase().includes(input.toLowerCase())}
-                options={[
-                  {
-                    value: Type.FullTime,
-                    label: 'Full-time',
-                  },
-                  {
-                    value: Type.Internship,
-                    label: 'Internship',
-                  },
-                  {
-                    value: Type.PartTime,
-                    label: 'Part-time',
-                  },
-                  {
-                    value: Type.Temporary,
-                    label: 'Temporary',
-                  },
-                ]}
-              />
-            </Form.Item>
-
-            <Form.Item wrapperCol={{ offset: 8, span: 16 }}>
-              <Button disabled={loading} type="primary" htmlType="submit">
-                Create
-              </Button>
-              <Link style={{ marginLeft: '16px' }} href={`/${AppRoutes.contracts}`} passHref>
-                <Button type="primary" danger>
-                  Cancel
-                </Button>
-              </Link>
-            </Form.Item>
-          </Space>
-        </Form>
+        <section>
+          <ContractForm
+            form={form}
+            button="Create"
+            company={company}
+            setCompany={setCompany}
+            employee={employee}
+            setEmployee={setEmployee}
+            file={file}
+            setFile={setFile}
+            isSending={loading}
+            onFinish={onFinish}
+          />
+        </section>
       </Space>
     </>
   );
